Make alb_no_logs resources conditional on create_alb
For all the resources in alb_no_logs that already use count to control
creation, add create_alb to the ternary.

Note that the condition has been reversed to !var.logging_enabled so
that it can be and-ed together with create_alb. As a result, the true
and false values on the right hand side of ternary have also been
reversed.