Provide an empty string as a default for some outputs
For outputs such as dns_name and load_balancer_id, add a list with an
empty string to the concat() list so that the empty string can be
returned as the output value (via element()). Otherwise, an error
occurs because it is not possible to select the 0-th element of an
empty list.